+++ apps/rslib-module/@mf-types/rslib-module/apis.d.ts
@@ -1,0 +1,4 @@
+export type RemoteKeys = 'rslib-module';
+type PackageType<T> = T extends 'rslib-module'
+  ? typeof import('rslib-module')
+  : any;