+++ packages/modernjs/src/cli/ssrPlugin.ts
@@ -7,6 +7,7 @@ import { ModuleFederationPlugin as RspackModuleFederationPlugin } from '@module-
 import UniverseEntryChunkTrackerPlugin from '@module-federation/node/universe-entry-chunk-tracker-plugin';
 import { updateStatsAndManifest } from './manifest';
 import { isDev } from './constant';
+import logger from './logger';
 
 export function setEnv() {
   process.env['MF_DISABLE_EMIT_STATS'] = 'true';
@@ -101,18 +102,13 @@ export const moduleFederationSSRPlugin = (
                         'Access-Control-Allow-Methods',
                         'GET, POST, PUT, DELETE, PATCH, OPTIONS',
                       );
-                      res.setHeader(
-                        'Access-Control-Allow-Headers',
-                        'X-Requested-With, content-type, Authorization',
-                      );
+                      res.setHeader('Access-Control-Allow-Headers', '*');
                       fs.createReadStream(filepath).pipe(res);
                     } else {
                       next();
                     }
                   } catch (err) {
-                    if (process.env.FEDERATION_DEBUG) {
-                      console.error(err);
-                    }
+                    logger.debug(err);
                     next();
                   }
                 },