+++ packages/retry-plugin/__tests__/retry.spec.ts
@@ -1,4 +1,3 @@
-import { scriptWithRetry } from '../src/script-retry';
 import { fetchWithRetry } from '../src/fetch-retry';
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 
@@ -22,62 +21,6 @@ vi.spyOn(document, 'createElement').mockImplementation(() => {
 });
 
 vi.spyOn(document.head, 'appendChild').mockImplementation(() => {});
-
-describe('scriptWithRetry', () => {
-  it('should create and append script element successfully', async () => {
-    const url = 'https://example.com/some-script.js';
-    const script = scriptWithRetry({
-      url,
-      retryTimes: 3,
-    });
-
-    mockScriptElement.onload();
-    expect(document.createElement).toHaveBeenCalledWith('script');
-    expect(script.src).toEqual(url);
-    expect(mockScriptElement.onload).toHaveBeenCalled();
-    expect(document.head.appendChild).toHaveBeenCalledTimes(0);
-  });
-
-  it('should retry loading script on error and eventually succeed', async () => {
-    const url = 'https://example.com/some-script.js';
-    const retryTimes = 2;
-
-    const script = scriptWithRetry({
-      url,
-      retryTimes,
-    });
-
-    // 前两次加载失败
-    mockScriptElement.onerror();
-    vi.advanceTimersByTime(2000 * retryTimes);
-
-    mockScriptElement.onerror();
-    vi.advanceTimersByTime(2000 * retryTimes);
-
-    mockScriptElement.onload();
-
-    // each call to appendChild represents one retry, so we can use the calledTimes of appendChild to verify the number of retries
-    expect(document.head.appendChild).toHaveBeenCalledTimes(retryTimes);
-  });
-
-  it('should stop retrying after maximum retries', async () => {
-    const url = 'https://example.com/some-script.js';
-    const retryTimes = 3;
-
-    const script = scriptWithRetry({
-      url,
-      retryTimes,
-    });
-
-    // all retries fail
-    for (let i = 0; i < retryTimes; i++) {
-      mockScriptElement.onerror();
-      vi.advanceTimersByTime(2000 * retryTimes);
-    }
-
-    expect(document.head.appendChild).toHaveBeenCalledTimes(retryTimes);
-  });
-});
 
 const mockGlobalFetch = (mockData) => {
   const mockFetch = vi.fn().mockResolvedValueOnce(mockResponse(200, mockData));
@@ -192,5 +135,23 @@ describe('fetchWithRetry', () => {
         retryDelay: 0,
       }),
     ).rejects.toThrow('Json parse error');
+  });
+
+  it('should build fallback URL from remote after retries fail', async () => {
+    mockErrorFetch();
+    const retryTimes = 3;
+    const responsePromise = fetchWithRetry({
+      url: 'https://example.com',
+      retryTimes,
+      retryDelay: 0,
+      fallback: (url) => `${url}/fallback`,
+    });
+    vi.advanceTimersByTime(2000 * retryTimes);
+
+    await expect(responsePromise).rejects.toThrow(
+      'The request failed three times and has now been abandoned',
+    );
+    expect(fetch).toHaveBeenCalledTimes(5); //first fetch + retryTimes fetch
+    expect(fetch).toHaveBeenLastCalledWith('https://example.com/fallback', {});
   });
 });