+++ apps/modernjs-ssr/remote-new-version/CHANGELOG.md
@@ -1,5 +1,24 @@
 # modernjs-ssr-remote-new-version
 
+## 0.1.36
+
+### Patch Changes
+
+- Updated dependencies [ad605d2]
+  - @module-federation/modern-js@0.6.14
+
+## 0.1.35
+
+### Patch Changes
+
+- @module-federation/modern-js@0.6.13
+
+## 0.1.34
+
+### Patch Changes
+
+- @module-federation/modern-js@0.6.12
+
 ## 0.1.33
 
 ### Patch Changes