+++ apps/modernjs/CHANGELOG.md
@@ -1,4 +1,35 @@
-# @module-federation/modernjs
+# @module-federation/modernjsapp
+
+## 0.1.62
+
+### Patch Changes
+
+- @module-federation/enhanced@0.6.16
+
+## 0.1.61
+
+### Patch Changes
+
+- @module-federation/enhanced@0.6.15
+
+## 0.1.60
+
+### Patch Changes
+
+- Updated dependencies [ad605d2]
+  - @module-federation/enhanced@0.6.14
+
+## 0.1.59
+
+### Patch Changes
+
+- @module-federation/enhanced@0.6.13
+
+## 0.1.58
+
+### Patch Changes
+
+- @module-federation/enhanced@0.6.12
 
 ## 0.1.57