+++ apps/router-demo/router-host-2000/CHANGELOG.md
@@ -1,5 +1,30 @@
 # host
 
+## 1.0.34
+
+### Patch Changes
+
+- Updated dependencies [ad605d2]
+  - @module-federation/bridge-react@0.6.14
+  - @module-federation/retry-plugin@0.6.14
+  - @module-federation/enhanced@0.6.14
+
+## 1.0.33
+
+### Patch Changes
+
+- @module-federation/enhanced@0.6.13
+- @module-federation/bridge-react@0.6.13
+- @module-federation/retry-plugin@0.6.13
+
+## 1.0.32
+
+### Patch Changes
+
+- @module-federation/enhanced@0.6.12
+- @module-federation/bridge-react@0.6.12
+- @module-federation/retry-plugin@0.6.12
+
 ## 1.0.31
 
 ### Patch Changes