+++ packages/bridge/bridge-react/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @module-federation/bridge-react
 
+## 0.6.16
+
+### Patch Changes
+
+- Updated dependencies [f779188]
+- Updated dependencies [024df60]
+  - @module-federation/sdk@0.6.16
+  - @module-federation/bridge-shared@0.6.16
+
+## 0.6.15
+
+### Patch Changes
+
+- @module-federation/sdk@0.6.15
+- @module-federation/bridge-shared@0.6.15
+
+## 0.6.14
+
+### Patch Changes
+
+- ad605d2: chore: unified logger
+- Updated dependencies [ad605d2]
+  - @module-federation/bridge-shared@0.6.14
+  - @module-federation/sdk@0.6.14
+
+## 0.6.13
+
+### Patch Changes
+
+- @module-federation/bridge-shared@0.6.13
+
+## 0.6.12
+
+### Patch Changes
+
+- @module-federation/bridge-shared@0.6.12
+
 ## 0.6.11
 
 ### Patch Changes