+++ packages/sdk/CHANGELOG.md
@@ -1,4 +1,23 @@
-# [1.1.0-canary.1](https://github.com/module-federation/core/compare/sdk-1.0.0...sdk-1.1.0-canary.1) (2023-12-05)
+# @module-federation/sdk
+
+## 0.6.16
+
+### Patch Changes
+
+- f779188: update rslog package version to resolve type error
+- 024df60: disable hoistTransitiveImports for better tree shake
+
+## 0.6.15
+
+## 0.6.14
+
+### Patch Changes
+
+- ad605d2: chore: unified logger
+
+## 0.6.13
+
+## 0.6.12
 
 ## 0.6.11