+++ packages/third-party-dts-extractor/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @module-federation/third-party-dts-extractor
 
+## 0.6.16
+
+## 0.6.15
+
+## 0.6.14
+
+## 0.6.13
+
+## 0.6.12
+
 ## 0.6.11
 
 ## 0.6.10